\name{print.fitEmax}
\alias{print.fitEmax}
\title{
Print a summary of the fitted Emax model
}
\description{
Print a summary of the fitted Emax model
}
\usage{
\method{print}{fitEmax}(x, ...)
}
\arguments{
  \item{x}{
Object output by \code{\link{fitEmax}}
}
  \item{\dots}{
No options implemented.
}
}
