\name{print.fitEmaxB}
\alias{print.fitEmaxB}
\title{
Print a summary of the fitted Bayesian Emax model
}
\description{
Print a summary of the fitted Bayesian Emax model
}
\usage{
\method{print}{fitEmaxB}(x, ...)
}
\arguments{
  \item{x}{
Object output by \code{\link{fitEmaxB}}
}
  \item{\dots}{
No options implemented.
}
}
