\name{summary.fitEmax}
\alias{summary.fitEmax}
\title{
Print a summary of the fitted Emax model
}
\description{
Print a summary of the fitted Emax model
}
\usage{
\method{summary}{fitEmax}(object, ...)
}
\arguments{
  \item{object}{
Object output by \code{\link{fitEmax}}
}
  \item{\dots}{
No options implemented.
}
}
