\name{summary.fitEmaxB}
\alias{summary.fitEmaxB}
\title{
Print a summary of the fitted Bayesian Emax model
}
\description{
Print a summary of the fitted Bayesian Emax model
}
\usage{
\method{summary}{fitEmaxB}(object, ...)
}
\arguments{
  \item{object}{
Object output by \code{\link{fitEmaxB}}
}
  \item{\dots}{
No options implemented.
}
}
