\name{showStanModels}
\alias{showStanModels}
\title{
Display \code{STAN} model code.
}
\description{
Display the \code{STAN} Bayesian model code for fitting Emax models
}
\usage{
showStanModels()
}

\author{
Neal Thomas
}

\seealso{
\code{\link{fitEmaxB}}, \code{\link{emaxsimB}}
}

\examples{
\dontrun{
showStanModels()
}
}

